clear bulk results on sim start, refactor bulk into separate files, update wow addon to generate bag data
