+++ tools/wowsimsexporter/WowSimsExporter/WowSimsExporter.lua
@@ -1,5 +1,8 @@
 -- Author      : generalwrex (Natop on Myzrael TBC)
 -- Create Date : 1/28/2022 9:30:08 AM
+--
+-- Update Date : 2023-04-04 Riotdog-GehennasEU: exporting bag items for bulk sim
+--
 
 WowSimsExporter = LibStub("AceAddon-3.0"):NewAddon("WowSimsExporter", "AceConsole-3.0", "AceEvent-3.0")
 
@@ -10,7 +13,7 @@ WowSimsExporter.Link = "https://wowsims.github.io/wotlk/"
 local AceGUI = LibStub("AceGUI-3.0")
 local LibParse = LibStub("LibParse")
 
-local version = "2.4"
+local version = "2.3"
 
 local defaults = {
 	profile = {
@@ -159,51 +162,58 @@ function WowSimsExporter:OpenWindow(input)
     end
 end
 
-function WowSimsExporter:GetGearEnchantGems(type)
-    local gear = {}
+function WowSimsExporter:createItemFromItemLink(itemLink)
+	local Id, Enchant, Gem1, Gem2, Gem3, Gem4 = self:ParseItemLink(itemLink)
+	item = {}
+	item.id = tonumber(Id)
+	item.enchant = tonumber(Enchant)
+	item.gems = {tonumber(Gem1), tonumber(Gem2), tonumber(Gem3), tonumber(Gem4)}
+	return item
+end
 
-	local slotNames = WowSimsExporter.slotNames
+function considerItemReplacement(itemLink)
+	-- TODO(Riotdog-GehennasEU): Is this sufficient? This seems to be what simc uses:
+	-- https://github.com/simulationcraft/simc-addon/blob/master/core.lua
+	-- Except we don't need the artifact check for wotlk classic.
+	-- We should probably filter some more, because this also returns e.g. the Mining Pick and items of the wrong armor type etc.. :D
+	return IsEquippableItem(itemLink)
+end
 
-    for slotNum = 1, #slotNames do
-        local slotId = GetInventorySlotInfo(slotNames[slotNum])
-        local itemLink = GetInventoryItemLink("player", slotId)
+function WowSimsExporter:GetGearEnchantGems(withBags)
+	self.Character.gear = {}
+	self.Character.glyphs = {}
+	self.Character.bagItems = {}
+	self.Character.items = nil
 
-        if itemLink then
-			gear[slotNum] = self:ParseItemLink(itemLink)
-        end
-    end
+	if not withBags then
+		local equippedGear = {}
+		local slotNames = WowSimsExporter.slotNames
+		for slotNum = 1, #slotNames do
+				local slotId = GetInventorySlotInfo(slotNames[slotNum])
+				local itemLink = GetInventoryItemLink("player", slotId)
+				if itemLink then
+					equippedGear[slotNum] = self:createItemFromItemLink(itemLink)
+				end
+		end
 
-    local bagSlots = {BANK_CONTAINER}
-    for bagNum = 0, NUM_BAG_SLOTS+NUM_BANKBAGSLOTS do
-        if GetContainerNumSlots(bagNum) > 0 then
-            table.insert(bagSlots, bagNum)
-        end
-    end
+		self.Character.spec = self:CheckCharacterSpec(self.Character.class)
+		self.Character.talents = self:CreateTalentEntry()
+		self:CreateGlyphEntry() -- wotlk
+		self:CreateProfessionEntry() -- wotlk
+		self.Character.gear.items = equippedGear
+		return self.Character
+	end
 
-    local bags = {}
-
-    for bagNum = 1, #bagSlots do
-        local slotCount = GetContainerNumSlots(bagSlots[bagNum])
-        for slotNum = 1, slotCount do
-            local _,_,_,_,_,_,itemLink = GetContainerItemInfo(bagSlots[bagNum], slotNum)
-            if itemLink and IsEquippableItem(itemLink) then
-                local _,_,_,_,_,_,_,_,itemEquipLoc,_,_ = GetItemInfo(itemLink)
-
-                if itemEquipLoc ~= "INVTYPE_BAG" then
-                    table.insert(bags, self:ParseItemLink(itemLink))
-                end
-            end
-        end
-    end
-
-	self.Character.spec = self:CheckCharacterSpec(self.Character.class)
-	self.Character.talents = self:CreateTalentEntry()
-	self:CreateGlyphEntry() -- wotlk
-	self:CreateProfessionEntry() -- wotlk
-	self.Character.gear.items = gear
-    self.Character.gear.bags = bags
-
-    return self.Character
+	local bagGear = {}
+	for bag = 0, 4 do
+		for slot = 1, GetContainerNumSlots(bag) do
+			local itemLink = GetContainerItemLink(bag, slot)
+			if itemLink and considerItemReplacement(itemLink) then
+				table.insert(bagGear, self:createItemFromItemLink(itemLink))
+			end
+		end
+	end
+  return {["items"] = bagGear}
 end
 
 
@@ -213,11 +223,7 @@ function WowSimsExporter:ParseItemLink(itemLink)
         itemLink,
         "|?c?f?f?(%x*)|?H?([^:]*):?(%d+):?(%d*):?(%d*):?(%d*):?(%d*):?(%d*):?(%-?%d*):?(%-?%d*):?(%d*):?(%d*):?(%-?%d*)|?h?%[?([^%[%]]*)%]?|?h?|?r?"
     )
-    local item = {}
-    item.id = tonumber(Id)
-    item.enchant = tonumber(Enchant)
-    item.gems = {tonumber(Gem1), tonumber(Gem2), tonumber(Gem3), tonumber(Gem4)}
-    return item
+    return Id, Enchant, Gem1, Gem2, Gem3, Gem4
 end
 
 function WowSimsExporter:OnInitialize()
@@ -316,9 +322,15 @@ function WowSimsExporter:CreateWindow(generate)
     jsonbox:SetFullHeight(true)
     jsonbox:DisableButton(true)
    
-	local function l_Generate()
-		WowSimsExporter.Character = WowSimsExporter:GetGearEnchantGems("player")
-		jsonbox:SetText(LibParse:JSONEncode(WowSimsExporter.Character)) 
+	local function l_Generate(withBags)
+		jsonbox:SetText('')
+		if not withBags then
+			WowSimsExporter.Character = WowSimsExporter:GetGearEnchantGems(withBags)
+			jsonbox:SetText(LibParse:JSONEncode(WowSimsExporter.Character))
+		else
+			local bagData = WowSimsExporter:GetGearEnchantGems(withBags)
+			jsonbox:SetText(LibParse:JSONEncode(bagData))
+		end
 		jsonbox:HighlightText()
 		jsonbox:SetFocus()
 
@@ -327,23 +339,28 @@ function WowSimsExporter:CreateWindow(generate)
 
 	if generate then l_Generate() end
 
-    local button = AceGUI:Create("Button")
-    button:SetText("Generate Data")
-    button:SetWidth(200)
+  local button = AceGUI:Create("Button")
+    button:SetText("Generate Data (Equipped Only)")
+    button:SetWidth(300)
 	button:SetCallback("OnClick", function()		
-		l_Generate()
+		l_Generate(false)
 	end)
-	
+
+	local extraButton = AceGUI:Create("Button")
+	extraButton:SetText("Generate Bulk Bag Data")
+	extraButton:SetWidth(300)
+	extraButton:SetCallback("OnClick", function()		
+		l_Generate(true)
+	end)
 	
 	local icon = AceGUI:Create("Icon")
 	icon:SetImage("Interface\\AddOns\\wowsimsexporter\\Skins\\wowsims.tga") 
 	icon:SetImageSize(32, 32)
 	icon:SetFullWidth(true)
 
-
-    local label = AceGUI:Create("Label")
+  local label = AceGUI:Create("Label")
 	label:SetFullWidth(true)
-    label:SetText([[
+  label:SetText([[
 
 To upload your character to the simuator, click on the url below that leads to the simuator website.
 
@@ -374,6 +391,7 @@ into the provided box and click "Import"
 		frame:AddChild(label)
 		WowSimsExporter:BuildLinks(frame, char)
 		frame:AddChild(button)
+		frame:AddChild(extraButton)
 		frame:AddChild(jsonbox)
 
 	end
