+++ ui/scss/core/individual_sim_ui/index.scss
@@ -11,6 +11,7 @@
 @import "../components/encounter_picker";
 @import "../components/exporters";
 @import "../components/enum_picker";
+@import "../components/bulk";
 @import "../components/gear_picker";
 @import "../components/icon_enum_picker";
 @import "../components/icon_picker";
